test(file): cover transfer command paths
